One more Decimal try extended

+++ hummingbot/core/data_type/order_book_row.py
@@ -8,8 +8,8 @@ class OrderBookRow(namedtuple("_OrderBookRow", "price, amount, update_id")):
     """
     Used to apply changes to OrderBook. OrderBook classes uses float internally for better performance over Decimal.
     """
-    price: float
-    amount: float
+    price: Decimal
+    amount: Decimal
     update_id: int
